Add files via upload
If a user already logged in, this will be the first page he/she will be redirected to. The buttons still don't work and idk how to auto insert the user's name, so that for example if I log in, it'll say "Hello Angelica!", it will also provide a preview of previous activity of the user.